Converts testing to not use pycds's builtin data
pycds's built in testing data will be deprecated, so it's time for
this to create its own. This patch modifies all of the test fixtures
to no longer import pycds.test_session.

This is a WIP. The tests no longer error out, but many still fail since
there now exist no testing data.