Minor improvements to the usage of sessions
There's some ugly hacks in here where session can be used either from
a DSN string that comes from app instantiation, or via a session
variable that is attached to the WSGI environment. The primary reason
for this is that when testing there are some weird cases were you need
to add data to a session, run the test case, and then roll back the
data. So the data are not accessible outside of the testing
transaction scope (say, if you make another connection using the
string-based connection parameters).

This patch adds a little logic to wrap the session into a simple
context manager, so that the session usage is consistent regardless of
wheter the session is coming from.